build: generate tests in out directory
