--- aliusnes/tests/65816.rs
@@ -1,1284 +1,0 @@
-mod tomharte;
-
-use tomharte::run_test;
-
-#[test]
-pub fn test_00() {
-	run_test("00.n");
-}
-
-#[test]
-pub fn test_01() {
-	run_test("01.n");
-}
-
-#[test]
-pub fn test_02() {
-	run_test("02.n");
-}
-
-#[test]
-pub fn test_03() {
-	run_test("03.n");
-}
-
-#[test]
-pub fn test_04() {
-	run_test("04.n");
-}
-
-#[test]
-pub fn test_05() {
-	run_test("05.n");
-}
-
-#[test]
-pub fn test_06() {
-	run_test("06.n");
-}
-
-#[test]
-pub fn test_07() {
-	run_test("07.n");
-}
-
-#[test]
-pub fn test_08() {
-	run_test("08.n");
-}
-
-#[test]
-pub fn test_09() {
-	run_test("09.n");
-}
-
-#[test]
-pub fn test_0a() {
-	run_test("0a.n");
-}
-
-#[test]
-pub fn test_0b() {
-	run_test("0b.n");
-}
-
-#[test]
-pub fn test_0c() {
-	run_test("0c.n");
-}
-
-#[test]
-pub fn test_0d() {
-	run_test("0d.n");
-}
-
-#[test]
-pub fn test_0e() {
-	run_test("0e.n");
-}
-
-#[test]
-pub fn test_0f() {
-	run_test("0f.n");
-}
-
-#[test]
-pub fn test_10() {
-	run_test("10.n");
-}
-
-#[test]
-pub fn test_11() {
-	run_test("11.n");
-}
-
-#[test]
-pub fn test_12() {
-	run_test("12.n");
-}
-
-#[test]
-pub fn test_13() {
-	run_test("13.n");
-}
-
-#[test]
-pub fn test_14() {
-	run_test("14.n");
-}
-
-#[test]
-pub fn test_15() {
-	run_test("15.n");
-}
-
-#[test]
-pub fn test_16() {
-	run_test("16.n");
-}
-
-#[test]
-pub fn test_17() {
-	run_test("17.n");
-}
-
-#[test]
-pub fn test_18() {
-	run_test("18.n");
-}
-
-#[test]
-pub fn test_19() {
-	run_test("19.n");
-}
-
-#[test]
-pub fn test_1a() {
-	run_test("1a.n");
-}
-
-#[test]
-pub fn test_1b() {
-	run_test("1b.n");
-}
-
-#[test]
-pub fn test_1c() {
-	run_test("1c.n");
-}
-
-#[test]
-pub fn test_1d() {
-	run_test("1d.n");
-}
-
-#[test]
-pub fn test_1e() {
-	run_test("1e.n");
-}
-
-#[test]
-pub fn test_1f() {
-	run_test("1f.n");
-}
-
-#[test]
-pub fn test_20() {
-	run_test("20.n");
-}
-
-#[test]
-pub fn test_21() {
-	run_test("21.n");
-}
-
-#[test]
-pub fn test_22() {
-	run_test("22.n");
-}
-
-#[test]
-pub fn test_23() {
-	run_test("23.n");
-}
-
-#[test]
-pub fn test_24() {
-	run_test("24.n");
-}
-
-#[test]
-pub fn test_25() {
-	run_test("25.n");
-}
-
-#[test]
-pub fn test_26() {
-	run_test("26.n");
-}
-
-#[test]
-pub fn test_27() {
-	run_test("27.n");
-}
-
-#[test]
-pub fn test_28() {
-	run_test("28.n");
-}
-
-#[test]
-pub fn test_29() {
-	run_test("29.n");
-}
-
-#[test]
-pub fn test_2a() {
-	run_test("2a.n");
-}
-
-#[test]
-pub fn test_2b() {
-	run_test("2b.n");
-}
-
-#[test]
-pub fn test_2c() {
-	run_test("2c.n");
-}
-
-#[test]
-pub fn test_2d() {
-	run_test("2d.n");
-}
-
-#[test]
-pub fn test_2e() {
-	run_test("2e.n");
-}
-
-#[test]
-pub fn test_2f() {
-	run_test("2f.n");
-}
-
-#[test]
-pub fn test_30() {
-	run_test("30.n");
-}
-
-#[test]
-pub fn test_31() {
-	run_test("31.n");
-}
-
-#[test]
-pub fn test_32() {
-	run_test("32.n");
-}
-
-#[test]
-pub fn test_33() {
-	run_test("33.n");
-}
-
-#[test]
-pub fn test_34() {
-	run_test("34.n");
-}
-
-#[test]
-pub fn test_35() {
-	run_test("35.n");
-}
-
-#[test]
-pub fn test_36() {
-	run_test("36.n");
-}
-
-#[test]
-pub fn test_37() {
-	run_test("37.n");
-}
-
-#[test]
-pub fn test_38() {
-	run_test("38.n");
-}
-
-#[test]
-pub fn test_39() {
-	run_test("39.n");
-}
-
-#[test]
-pub fn test_3a() {
-	run_test("3a.n");
-}
-
-#[test]
-pub fn test_3b() {
-	run_test("3b.n");
-}
-
-#[test]
-pub fn test_3c() {
-	run_test("3c.n");
-}
-
-#[test]
-pub fn test_3d() {
-	run_test("3d.n");
-}
-
-#[test]
-pub fn test_3e() {
-	run_test("3e.n");
-}
-
-#[test]
-pub fn test_3f() {
-	run_test("3f.n");
-}
-
-#[test]
-pub fn test_40() {
-	run_test("40.n");
-}
-
-#[test]
-pub fn test_41() {
-	run_test("41.n");
-}
-
-#[test]
-pub fn test_42() {
-	run_test("42.n");
-}
-
-#[test]
-pub fn test_43() {
-	run_test("43.n");
-}
-
-#[test]
-pub fn test_44() {
-	run_test("44.n");
-}
-
-#[test]
-pub fn test_45() {
-	run_test("45.n");
-}
-
-#[test]
-pub fn test_46() {
-	run_test("46.n");
-}
-
-#[test]
-pub fn test_47() {
-	run_test("47.n");
-}
-
-#[test]
-pub fn test_48() {
-	run_test("48.n");
-}
-
-#[test]
-pub fn test_49() {
-	run_test("49.n");
-}
-
-#[test]
-pub fn test_4a() {
-	run_test("4a.n");
-}
-
-#[test]
-pub fn test_4b() {
-	run_test("4b.n");
-}
-
-#[test]
-pub fn test_4c() {
-	run_test("4c.n");
-}
-
-#[test]
-pub fn test_4d() {
-	run_test("4d.n");
-}
-
-#[test]
-pub fn test_4e() {
-	run_test("4e.n");
-}
-
-#[test]
-pub fn test_4f() {
-	run_test("4f.n");
-}
-
-#[test]
-pub fn test_50() {
-	run_test("50.n");
-}
-
-#[test]
-pub fn test_51() {
-	run_test("51.n");
-}
-
-#[test]
-pub fn test_52() {
-	run_test("52.n");
-}
-
-#[test]
-pub fn test_53() {
-	run_test("53.n");
-}
-
-#[test]
-pub fn test_54() {
-	run_test("54.n");
-}
-
-#[test]
-pub fn test_55() {
-	run_test("55.n");
-}
-
-#[test]
-pub fn test_56() {
-	run_test("56.n");
-}
-
-#[test]
-pub fn test_57() {
-	run_test("57.n");
-}
-
-#[test]
-pub fn test_58() {
-	run_test("58.n");
-}
-
-#[test]
-pub fn test_59() {
-	run_test("59.n");
-}
-
-#[test]
-pub fn test_5a() {
-	run_test("5a.n");
-}
-
-#[test]
-pub fn test_5b() {
-	run_test("5b.n");
-}
-
-#[test]
-pub fn test_5c() {
-	run_test("5c.n");
-}
-
-#[test]
-pub fn test_5d() {
-	run_test("5d.n");
-}
-
-#[test]
-pub fn test_5e() {
-	run_test("5e.n");
-}
-
-#[test]
-pub fn test_5f() {
-	run_test("5f.n");
-}
-
-#[test]
-pub fn test_60() {
-	run_test("60.n");
-}
-
-#[test]
-pub fn test_61() {
-	run_test("61.n");
-}
-
-#[test]
-pub fn test_62() {
-	run_test("62.n");
-}
-
-#[test]
-pub fn test_63() {
-	run_test("63.n");
-}
-
-#[test]
-pub fn test_64() {
-	run_test("64.n");
-}
-
-#[test]
-pub fn test_65() {
-	run_test("65.n");
-}
-
-#[test]
-pub fn test_66() {
-	run_test("66.n");
-}
-
-#[test]
-pub fn test_67() {
-	run_test("67.n");
-}
-
-#[test]
-pub fn test_68() {
-	run_test("68.n");
-}
-
-#[test]
-pub fn test_69() {
-	run_test("69.n");
-}
-
-#[test]
-pub fn test_6a() {
-	run_test("6a.n");
-}
-
-#[test]
-pub fn test_6b() {
-	run_test("6b.n");
-}
-
-#[test]
-pub fn test_6c() {
-	run_test("6c.n");
-}
-
-#[test]
-pub fn test_6d() {
-	run_test("6d.n");
-}
-
-#[test]
-pub fn test_6e() {
-	run_test("6e.n");
-}
-
-#[test]
-pub fn test_6f() {
-	run_test("6f.n");
-}
-
-#[test]
-pub fn test_70() {
-	run_test("70.n");
-}
-
-#[test]
-pub fn test_71() {
-	run_test("71.n");
-}
-
-#[test]
-pub fn test_72() {
-	run_test("72.n");
-}
-
-#[test]
-pub fn test_73() {
-	run_test("73.n");
-}
-
-#[test]
-pub fn test_74() {
-	run_test("74.n");
-}
-
-#[test]
-pub fn test_75() {
-	run_test("75.n");
-}
-
-#[test]
-pub fn test_76() {
-	run_test("76.n");
-}
-
-#[test]
-pub fn test_77() {
-	run_test("77.n");
-}
-
-#[test]
-pub fn test_78() {
-	run_test("78.n");
-}
-
-#[test]
-pub fn test_79() {
-	run_test("79.n");
-}
-
-#[test]
-pub fn test_7a() {
-	run_test("7a.n");
-}
-
-#[test]
-pub fn test_7b() {
-	run_test("7b.n");
-}
-
-#[test]
-pub fn test_7c() {
-	run_test("7c.n");
-}
-
-#[test]
-pub fn test_7d() {
-	run_test("7d.n");
-}
-
-#[test]
-pub fn test_7e() {
-	run_test("7e.n");
-}
-
-#[test]
-pub fn test_7f() {
-	run_test("7f.n");
-}
-
-#[test]
-pub fn test_80() {
-	run_test("80.n");
-}
-
-#[test]
-pub fn test_81() {
-	run_test("81.n");
-}
-
-#[test]
-pub fn test_82() {
-	run_test("82.n");
-}
-
-#[test]
-pub fn test_83() {
-	run_test("83.n");
-}
-
-#[test]
-pub fn test_84() {
-	run_test("84.n");
-}
-
-#[test]
-pub fn test_85() {
-	run_test("85.n");
-}
-
-#[test]
-pub fn test_86() {
-	run_test("86.n");
-}
-
-#[test]
-pub fn test_87() {
-	run_test("87.n");
-}
-
-#[test]
-pub fn test_88() {
-	run_test("88.n");
-}
-
-#[test]
-pub fn test_89() {
-	run_test("89.n");
-}
-
-#[test]
-pub fn test_8a() {
-	run_test("8a.n");
-}
-
-#[test]
-pub fn test_8b() {
-	run_test("8b.n");
-}
-
-#[test]
-pub fn test_8c() {
-	run_test("8c.n");
-}
-
-#[test]
-pub fn test_8d() {
-	run_test("8d.n");
-}
-
-#[test]
-pub fn test_8e() {
-	run_test("8e.n");
-}
-
-#[test]
-pub fn test_8f() {
-	run_test("8f.n");
-}
-
-#[test]
-pub fn test_90() {
-	run_test("90.n");
-}
-
-#[test]
-pub fn test_91() {
-	run_test("91.n");
-}
-
-#[test]
-pub fn test_92() {
-	run_test("92.n");
-}
-
-#[test]
-pub fn test_93() {
-	run_test("93.n");
-}
-
-#[test]
-pub fn test_94() {
-	run_test("94.n");
-}
-
-#[test]
-pub fn test_95() {
-	run_test("95.n");
-}
-
-#[test]
-pub fn test_96() {
-	run_test("96.n");
-}
-
-#[test]
-pub fn test_97() {
-	run_test("97.n");
-}
-
-#[test]
-pub fn test_98() {
-	run_test("98.n");
-}
-
-#[test]
-pub fn test_99() {
-	run_test("99.n");
-}
-
-#[test]
-pub fn test_9a() {
-	run_test("9a.n");
-}
-
-#[test]
-pub fn test_9b() {
-	run_test("9b.n");
-}
-
-#[test]
-pub fn test_9c() {
-	run_test("9c.n");
-}
-
-#[test]
-pub fn test_9d() {
-	run_test("9d.n");
-}
-
-#[test]
-pub fn test_9e() {
-	run_test("9e.n");
-}
-
-#[test]
-pub fn test_9f() {
-	run_test("9f.n");
-}
-
-#[test]
-pub fn test_a0() {
-	run_test("a0.n");
-}
-
-#[test]
-pub fn test_a1() {
-	run_test("a1.n");
-}
-
-#[test]
-pub fn test_a2() {
-	run_test("a2.n");
-}
-
-#[test]
-pub fn test_a3() {
-	run_test("a3.n");
-}
-
-#[test]
-pub fn test_a4() {
-	run_test("a4.n");
-}
-
-#[test]
-pub fn test_a5() {
-	run_test("a5.n");
-}
-
-#[test]
-pub fn test_a6() {
-	run_test("a6.n");
-}
-
-#[test]
-pub fn test_a7() {
-	run_test("a7.n");
-}
-
-#[test]
-pub fn test_a8() {
-	run_test("a8.n");
-}
-
-#[test]
-pub fn test_a9() {
-	run_test("a9.n");
-}
-
-#[test]
-pub fn test_aa() {
-	run_test("aa.n");
-}
-
-#[test]
-pub fn test_ab() {
-	run_test("ab.n");
-}
-
-#[test]
-pub fn test_ac() {
-	run_test("ac.n");
-}
-
-#[test]
-pub fn test_ad() {
-	run_test("ad.n");
-}
-
-#[test]
-pub fn test_ae() {
-	run_test("ae.n");
-}
-
-#[test]
-pub fn test_af() {
-	run_test("af.n");
-}
-
-#[test]
-pub fn test_b0() {
-	run_test("b0.n");
-}
-
-#[test]
-pub fn test_b1() {
-	run_test("b1.n");
-}
-
-#[test]
-pub fn test_b2() {
-	run_test("b2.n");
-}
-
-#[test]
-pub fn test_b3() {
-	run_test("b3.n");
-}
-
-#[test]
-pub fn test_b4() {
-	run_test("b4.n");
-}
-
-#[test]
-pub fn test_b5() {
-	run_test("b5.n");
-}
-
-#[test]
-pub fn test_b6() {
-	run_test("b6.n");
-}
-
-#[test]
-pub fn test_b7() {
-	run_test("b7.n");
-}
-
-#[test]
-pub fn test_b8() {
-	run_test("b8.n");
-}
-
-#[test]
-pub fn test_b9() {
-	run_test("b9.n");
-}
-
-#[test]
-pub fn test_ba() {
-	run_test("ba.n");
-}
-
-#[test]
-pub fn test_bb() {
-	run_test("bb.n");
-}
-
-#[test]
-pub fn test_bc() {
-	run_test("bc.n");
-}
-
-#[test]
-pub fn test_bd() {
-	run_test("bd.n");
-}
-
-#[test]
-pub fn test_be() {
-	run_test("be.n");
-}
-
-#[test]
-pub fn test_bf() {
-	run_test("bf.n");
-}
-
-#[test]
-pub fn test_c0() {
-	run_test("c0.n");
-}
-
-#[test]
-pub fn test_c1() {
-	run_test("c1.n");
-}
-
-#[test]
-pub fn test_c2() {
-	run_test("c2.n");
-}
-
-#[test]
-pub fn test_c3() {
-	run_test("c3.n");
-}
-
-#[test]
-pub fn test_c4() {
-	run_test("c4.n");
-}
-
-#[test]
-pub fn test_c5() {
-	run_test("c5.n");
-}
-
-#[test]
-pub fn test_c6() {
-	run_test("c6.n");
-}
-
-#[test]
-pub fn test_c7() {
-	run_test("c7.n");
-}
-
-#[test]
-pub fn test_c8() {
-	run_test("c8.n");
-}
-
-#[test]
-pub fn test_c9() {
-	run_test("c9.n");
-}
-
-#[test]
-pub fn test_ca() {
-	run_test("ca.n");
-}
-
-#[test]
-pub fn test_cb() {
-	run_test("cb.n");
-}
-
-#[test]
-pub fn test_cc() {
-	run_test("cc.n");
-}
-
-#[test]
-pub fn test_cd() {
-	run_test("cd.n");
-}
-
-#[test]
-pub fn test_ce() {
-	run_test("ce.n");
-}
-
-#[test]
-pub fn test_cf() {
-	run_test("cf.n");
-}
-
-#[test]
-pub fn test_d0() {
-	run_test("d0.n");
-}
-
-#[test]
-pub fn test_d1() {
-	run_test("d1.n");
-}
-
-#[test]
-pub fn test_d2() {
-	run_test("d2.n");
-}
-
-#[test]
-pub fn test_d3() {
-	run_test("d3.n");
-}
-
-#[test]
-pub fn test_d4() {
-	run_test("d4.n");
-}
-
-#[test]
-pub fn test_d5() {
-	run_test("d5.n");
-}
-
-#[test]
-pub fn test_d6() {
-	run_test("d6.n");
-}
-
-#[test]
-pub fn test_d7() {
-	run_test("d7.n");
-}
-
-#[test]
-pub fn test_d8() {
-	run_test("d8.n");
-}
-
-#[test]
-pub fn test_d9() {
-	run_test("d9.n");
-}
-
-#[test]
-pub fn test_da() {
-	run_test("da.n");
-}
-
-#[test]
-pub fn test_db() {
-	run_test("db.n");
-}
-
-#[test]
-pub fn test_dc() {
-	run_test("dc.n");
-}
-
-#[test]
-pub fn test_dd() {
-	run_test("dd.n");
-}
-
-#[test]
-pub fn test_de() {
-	run_test("de.n");
-}
-
-#[test]
-pub fn test_df() {
-	run_test("df.n");
-}
-
-#[test]
-pub fn test_e0() {
-	run_test("e0.n");
-}
-
-#[test]
-pub fn test_e1() {
-	run_test("e1.n");
-}
-
-#[test]
-pub fn test_e2() {
-	run_test("e2.n");
-}
-
-#[test]
-pub fn test_e3() {
-	run_test("e3.n");
-}
-
-#[test]
-pub fn test_e4() {
-	run_test("e4.n");
-}
-
-#[test]
-pub fn test_e5() {
-	run_test("e5.n");
-}
-
-#[test]
-pub fn test_e6() {
-	run_test("e6.n");
-}
-
-#[test]
-pub fn test_e7() {
-	run_test("e7.n");
-}
-
-#[test]
-pub fn test_e8() {
-	run_test("e8.n");
-}
-
-#[test]
-pub fn test_e9() {
-	run_test("e9.n");
-}
-
-#[test]
-pub fn test_ea() {
-	run_test("ea.n");
-}
-
-#[test]
-pub fn test_eb() {
-	run_test("eb.n");
-}
-
-#[test]
-pub fn test_ec() {
-	run_test("ec.n");
-}
-
-#[test]
-pub fn test_ed() {
-	run_test("ed.n");
-}
-
-#[test]
-pub fn test_ee() {
-	run_test("ee.n");
-}
-
-#[test]
-pub fn test_ef() {
-	run_test("ef.n");
-}
-
-#[test]
-pub fn test_f0() {
-	run_test("f0.n");
-}
-
-#[test]
-pub fn test_f1() {
-	run_test("f1.n");
-}
-
-#[test]
-pub fn test_f2() {
-	run_test("f2.n");
-}
-
-#[test]
-pub fn test_f3() {
-	run_test("f3.n");
-}
-
-#[test]
-pub fn test_f4() {
-	run_test("f4.n");
-}
-
-#[test]
-pub fn test_f5() {
-	run_test("f5.n");
-}
-
-#[test]
-pub fn test_f6() {
-	run_test("f6.n");
-}
-
-#[test]
-pub fn test_f7() {
-	run_test("f7.n");
-}
-
-#[test]
-pub fn test_f8() {
-	run_test("f8.n");
-}
-
-#[test]
-pub fn test_f9() {
-	run_test("f9.n");
-}
-
-#[test]
-pub fn test_fa() {
-	run_test("fa.n");
-}
-
-#[test]
-pub fn test_fb() {
-	run_test("fb.n");
-}
-
-#[test]
-pub fn test_fc() {
-	run_test("fc.n");
-}
-
-#[test]
-pub fn test_fd() {
-	run_test("fd.n");
-}
-
-#[test]
-pub fn test_fe() {
-	run_test("fe.n");
-}
-
-#[test]
-pub fn test_ff() {
-	run_test("ff.n");
-}
-
